+++ OperationEncoderBin.java
@@ -5,10 +5,10 @@ public class OperationEncoderBin implements OperationEncoder, OperationBinConst 
 	public byte[] encode(Operation op) throws Exception {
 		ByteArrayOutputStream buf = new ByteArrayOutputStream();
 		DataOutputStream out = new DataOutputStream(buf);
-		out.writeInt(op.totalLength);
-		out.writeLong(op.ID);
-		out.writeInt(op.opCode);
-		out.writeInt(op.numberOperands);
+		out.writeByte(op.totalLength);
+		out.writeByte(op.ID);
+		out.writeByte(op.opCode);
+		out.writeByte(op.numberOperands);
 		out.writeShort(op.operandOne);
 		out.writeShort(op.operandTwo);
